Second tree displayed
